#13 - remove images
