+++ src/api/faculties.ts
@@ -2,16 +2,6 @@ import { useFetch } from "../utils/reactQuery.utils"
 import { apiRoutes } from "../constants/api"
 import { pathToUrl } from "../utils/routes.utils"
 
-type FacultyProps = {
-  id: number
-  externalId: number
-  name: string
-}
-
-type FacultiesProps = {
-  data: Array<FacultyProps>
-}
-
 export const useGetFaculties = () =>
   useFetch<FacultiesProps>(apiRoutes.getFaculties)
