+++ src/components/Card/Card.tsx
@@ -5,11 +5,11 @@ interface CardProps {
   name: string
   path: string
   className?: string
-  image?: string
 }
 
 export const Card = (props: CardProps) => {
-  const { name, path, className = "bg-gray-400", image } = props
+  const { name, path, className = "bg-gray-400" } = props
+
   return (
     <Link to={path}>
       <div
@@ -18,13 +18,6 @@ export const Card = (props: CardProps) => {
           className
         )}
       >
-        {image && (
-          <img
-            src={image}
-            alt={name}
-            className="col-end-1 row-end-1 brightness-75"
-          />
-        )}
         <p className="z-10 col-end-1 row-end-1 m-auto px-9 text-center leading-relaxed">
           {name}
         </p>
