+++ src/components/Group/Group.tsx
@@ -1,4 +1,5 @@
 import classNames from "classnames"
+import { colors } from "../../constants/colors"
 import { Card } from "../Card"
 
 type ElementProps = {
@@ -10,11 +11,11 @@ interface GroupProps {
   title: string
   elements: Array<ElementProps>
   route: (id: number) => string
-  images?: Array<string>
 }
 
 export const Group = (props: GroupProps) => {
-  const { title, elements, route, images } = props
+  const { title, elements, route } = props
+
   return (
     <div className="px-20 py-10">
       <h2 className="text-center text-[45px] font-semibold">{title}</h2>
@@ -31,7 +32,7 @@ export const Group = (props: GroupProps) => {
             key={id}
             name={name}
             path={route(id)}
-            image={images && images[index]}
+            className={colors[index % colors.length]}
           />
         ))}
       </div>
