+++ src/constants/colors.ts
@@ -1,0 +1,10 @@
+export const colors: Array<string> = [
+  "bg-red-500",
+  "bg-blue-500",
+  "bg-green-500",
+  "bg-purple-500",
+  "bg-yellow-500",
+  "bg-lime-500",
+  "bg-sky-500",
+  "bg-orange-500"
+]
