+++ src/types/types.ts
@@ -1,0 +1,10 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
+type FacultyProps = {
+  id: number
+  externalId: number
+  name: string
+}
+
+type FacultiesProps = {
+  data: Array<FacultyProps>
+}
